refactor(backend): extract stuff to module
